Fix duration format and type

+++ app/src/main/java/edu/unipi/students/omadara/musicplayer/Album.java
@@ -7,7 +7,7 @@ public class Album {
     private String title;
     private String artist;
     private int trackCount;
-    private int duration;
+    private long duration;
     private Bitmap thumbnail;
 
     public String getId() {
@@ -42,11 +42,11 @@ public class Album {
         this.trackCount = trackCount;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
 
-    public void setDuration(int duration) {
+    public void setDuration(long duration) {
         this.duration = duration;
     }
